+++ src/driver/driver.rs
@@ -8,6 +8,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+#![no_start]
+
 #[cfg(rustdoc)]
 extern crate "rustdoc" as this;
